@@ -1,18 +1,19 @@
 import 'package:flutter/material.dart';
-import 'package:magazapp_flutter/components/reusable_card.dart';
 import 'package:magazapp_flutter/main.dart';
-import 'package:magazapp_flutter/screens/home_screen.dart';
-import 'package:magazapp_flutter/screens/product_page.dart';
 
 class SummaryPage extends StatefulWidget {
-
   final int id;
   final String title;
   final String image;
   final double price;
   final int qty;
 
-  SummaryPage({this.id, this.title, this.image, this.price, this.qty});
+  SummaryPage(
+      {required this.id,
+      required this.title,
+      required this.image,
+      required this.price,
+      required this.qty});
 
   @override
   _SummaryPageState createState() => _SummaryPageState();
@@ -37,84 +38,18 @@ class _SummaryPageState extends State<SummaryPage> {
                   ),
                 ),
               ),
-              Column(
-                children: List.generate(
-                  cartList.length,
-                      (index) => ReusableCard(
-                    color: Colors.grey.shade300,
-                    cardChild: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      mainAxisAlignment: MainAxisAlignment.start,
-                      children: <Widget>[
-                        Row(
-                          children: <Widget>[
-                            ReusableCard(
-                              cardHeigth: 75.0,
-                              cardWidth: 75.0,
-                              cardChild: Container(
-                                decoration: BoxDecoration(
-                                  borderRadius: BorderRadius.circular(10.0),
-                                  image: DecorationImage(
-                                    image:
-                                    AssetImage(cartList[index].image),
-                                    fit: BoxFit.cover,
-                                  ),
-                                ),
-                              ),
-                            ),
-                            Container(
-                              child: Column(
-                                crossAxisAlignment: CrossAxisAlignment.start,
-                                children: <Widget>[
-                                  SizedBox(
-                                    height: 5.0,
-                                  ),
-                                  Text(
-                                    cartList[index].title,
-                                    style: TextStyle(
-                                      fontSize: 15.0,
-                                      fontWeight: FontWeight.bold,
-                                      color: Colors.black,
-                                    ),
-                                  ),
-                                  SizedBox(
-                                    height: 2.0,
-                                  ),
-                                  Text(
-                                    "Fiyat : ${(cartList[index].price * cartList[index].qty).toStringAsFixed(2)} TL",
-                                    style: TextStyle(
-                                      fontSize: 13.0,
-                                    ),
-                                  ),
-                                  SizedBox(
-                                    height: 2.0,
-                                  ),
-                                  Text(
-                                    'Adet : ${cartList[index].qty}',
-                                    style: TextStyle(
-                                      fontWeight: FontWeight.bold,
-                                      fontSize: 15.0,
-                                      color: Colors.black,
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            ),
-                          ],
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-              ),
               SizedBox(
                 height: 20.0,
               ),
               GestureDetector(
                 onTap: () {
-                  Navigator.push(context, MaterialPageRoute(
-                    builder: (_) => MainPage(),
-                  ),
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (_) => MainPage(
+                        title: '',
+                      ),
+                    ),
                   );
                 },
                 child: Container(
