@@ -6,10 +6,10 @@ class OrderList {
   int qty;
 
   OrderList({
-    this.id,
-    this.title,
-    this.price,
-    this.image,
-    this.qty,
+    required this.id,
+    required this.title,
+    required this.price,
+    required this.image,
+    required this.qty,
   });
 }
