@@ -11,8 +11,6 @@ final double fontSize = 20.0;
 final double cardWidth = 175.0;
 final double cardHeigth = 175.0;
 
-
-
 class _CategoryScreenState extends State<CategoryScreen> {
   @override
   Widget build(BuildContext context) {
@@ -26,20 +24,12 @@ class _CategoryScreenState extends State<CategoryScreen> {
             children: List.generate(
               categories.length,
               (index) => CategoryCard(
-                onPress: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (_) => categories[index]["category_route"],
-                    ),
-                  );
-                },
-                fontSize: fontSize,
-                cardWidth: cardWidth,
-                cardHeigth: cardHeigth,
-                categoryTitle: "${categories[index]["title"]}",
-                assetImage: AssetImage(categories[index]["image"]),
-              ),
+                  color: Colors.white,
+                  assetImage: AssetImage(categories[index]["image"]),
+                  categoryTitle: categories[index]["title"],
+                  fontSize: fontSize,
+                  cardWidth: cardWidth,
+                  cardHeigth: cardHeigth),
             ),
           ),
         ),
