@@ -1,13 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:magazapp_flutter/components/reusable_card.dart';
 import 'package:magazapp_flutter/screens/product_page.dart';
 import 'package:magazapp_flutter/products/product.dart';
-import 'package:magazapp_flutter/settings_screens/my_orders.dart';
 import 'package:magazapp_flutter/settings_screens/order_track.dart';
 import 'package:magazapp_flutter/settings_screens/returns.dart';
 import 'package:magazapp_flutter/settings_screens/settings_page.dart';
 import 'package:magazapp_flutter/settings_screens/user_profile.dart';
-import 'package:magazapp_flutter/settings_screens/wishlist.dart';
 
 class ProfileScreen extends StatefulWidget {
   @override
@@ -28,7 +25,8 @@ class _ProfileScreenState extends State<ProfileScreen> {
             crossAxisAlignment: CrossAxisAlignment.stretch,
             children: <Widget>[
               Padding(
-                padding: const EdgeInsets.only(left: 5.0, top: 5.0, bottom: 5.0),
+                padding:
+                    const EdgeInsets.only(left: 5.0, top: 5.0, bottom: 5.0),
                 child: Text(
                   "Son Görüntülenen Ürünler:",
                   style: TextStyle(
@@ -40,66 +38,66 @@ class _ProfileScreenState extends State<ProfileScreen> {
               SingleChildScrollView(
                 scrollDirection: Axis.horizontal,
                 child: Row(
-                  children: List.generate(lastOrders.length, (index) =>
-                      ReusableCard(
-                        onPress: () {
-                          Navigator.push(
-                            context,
-                            MaterialPageRoute(
-                              builder: (_) => ProductPage(
-                                id: lastOrders[index]["id"],
-                                title: lastOrders[index]["title"],
-                                price: lastOrders[index]["price"],
-                                image: lastOrders[index]["image"],
-                              ),
+                  children: List.generate(
+                    lastOrders.length,
+                    (index) => GestureDetector(
+                      onTap: () {
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                            builder: (_) => ProductPage(
+                              id: lastOrders[index]["id"],
+                              title: lastOrders[index]["title"],
+                              price: lastOrders[index]["price"],
+                              image: lastOrders[index]["image"],
                             ),
-                          );
-                        },
-                        cardHeigth: 150.0,
-                        cardWidth: 150.0,
-                        cardChild: Container(
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(10.0),
-                            image: DecorationImage(
-                                image: AssetImage(lastOrders[index]["image"]),
-                                fit: BoxFit.cover),
                           ),
+                        );
+                      },
+                      child: Container(
+                        height: 150.0,
+                        width: 150.0,
+                        decoration: BoxDecoration(
+                          borderRadius: BorderRadius.circular(10.0),
+                          image: DecorationImage(
+                              image: AssetImage(lastOrders[index]["image"]),
+                              fit: BoxFit.cover),
                         ),
                       ),
-                  ),
-                ),
-              ),
-              ReusableCard(
-                onPress: () {
-                  Navigator.push(context, MaterialPageRoute(builder: (_) => UserProfile(),),);
-                },
-                color: Colors.blueGrey.shade100,
-                cardHeigth: 50.0,
-                cardWidth: 250.0,
-                cardChild: Center(
-                  child: Text(
-                    "Hesabım",
-                    style: TextStyle(
-                      color: Colors.black87,
-                      fontSize: 18.0,
-                      fontWeight: FontWeight.bold,
                     ),
                   ),
                 ),
               ),
-              ReusableCard(
-                onPress: () {
+              GestureDetector(
+                onTap: () {
                   Navigator.push(
                     context,
                     MaterialPageRoute(
-                      builder: (_) => MyOrders(),
+                      builder: (_) => UserProfile(),
                     ),
                   );
                 },
+                child: Container(
+                  color: Colors.blueGrey.shade100,
+                  height: 50.0,
+                  width: 250.0,
+                  child: Center(
+                    child: Text(
+                      "Hesabım",
+                      style: TextStyle(
+                        color: Colors.black87,
+                        fontSize: 18.0,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                  ),
+                ),
+              ),
+              Container(
                 color: Colors.blueGrey.shade100,
-                cardHeigth: 50.0,
-                cardWidth: 250.0,
-                cardChild: Center(
+                height: 50.0,
+                width: 250.0,
+                child: Center(
                   child: Text(
                     "Siparişlerim",
                     style: TextStyle(
@@ -110,32 +108,36 @@ class _ProfileScreenState extends State<ProfileScreen> {
                   ),
                 ),
               ),
-              ReusableCard(
-                onPress: () {
-                  Navigator.push(context, MaterialPageRoute(builder: (_) => OrderTrack(),),);
+              GestureDetector(
+                onTap: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (_) => OrderTrack(),
+                    ),
+                  );
                 },
-                color: Colors.blueGrey.shade100,
-                cardHeigth: 50.0,
-                cardWidth: 250.0,
-                cardChild: Center(
-                  child: Text(
-                    "Kargolarım",
-                    style: TextStyle(
-                      color: Colors.black87,
-                      fontSize: 18.0,
-                      fontWeight: FontWeight.bold,
+                child: Container(
+                  color: Colors.blueGrey.shade100,
+                  height: 50.0,
+                  width: 250.0,
+                  child: Center(
+                    child: Text(
+                      "Kargolarım",
+                      style: TextStyle(
+                        color: Colors.black87,
+                        fontSize: 18.0,
+                        fontWeight: FontWeight.bold,
+                      ),
                     ),
                   ),
                 ),
               ),
-              ReusableCard(
-                onPress: () {
-                  Navigator.push(context, MaterialPageRoute(builder: (_) => WishList(),),);
-                },
+              Container(
                 color: Colors.blueGrey.shade100,
-                cardHeigth: 50.0,
-                cardWidth: 250.0,
-                cardChild: Center(
+                height: 50.0,
+                width: 250.0,
+                child: Center(
                   child: Text(
                     "İstek Listem",
                     style: TextStyle(
@@ -146,38 +148,52 @@ class _ProfileScreenState extends State<ProfileScreen> {
                   ),
                 ),
               ),
-              ReusableCard(
-                onPress: () {
-                  Navigator.push(context, MaterialPageRoute(builder: (_) => Returns(),),);
+              GestureDetector(
+                onTap: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (_) => Returns(),
+                    ),
+                  );
                 },
-                color: Colors.blueGrey.shade100,
-                cardHeigth: 50.0,
-                cardWidth: 250.0,
-                cardChild: Center(
-                  child: Text(
-                    "İadelerim",
-                    style: TextStyle(
-                      color: Colors.black87,
-                      fontSize: 18.0,
-                      fontWeight: FontWeight.bold,
+                child: Container(
+                  color: Colors.blueGrey.shade100,
+                  height: 50.0,
+                  width: 250.0,
+                  child: Center(
+                    child: Text(
+                      "İadelerim",
+                      style: TextStyle(
+                        color: Colors.black87,
+                        fontSize: 18.0,
+                        fontWeight: FontWeight.bold,
+                      ),
                     ),
                   ),
                 ),
               ),
-              ReusableCard(
-                onPress: () {
-                  Navigator.push(context, MaterialPageRoute(builder: (_) => SettingsPage(),),);
+              GestureDetector(
+                onTap: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (_) => SettingsPage(),
+                    ),
+                  );
                 },
-                color: Colors.blueGrey.shade100,
-                cardHeigth: 50.0,
-                cardWidth: 250.0,
-                cardChild: Center(
-                  child: Text(
-                    "Ayarlar",
-                    style: TextStyle(
-                      color: Colors.black87,
-                      fontSize: 18.0,
-                      fontWeight: FontWeight.bold,
+                child: Container(
+                  color: Colors.blueGrey.shade100,
+                  height: 50.0,
+                  width: 250.0,
+                  child: Center(
+                    child: Text(
+                      "Ayarlar",
+                      style: TextStyle(
+                        color: Colors.black87,
+                        fontSize: 18.0,
+                        fontWeight: FontWeight.bold,
+                      ),
                     ),
                   ),
                 ),
